fix: abort activation scriipts if secrets are missing

+++ modules/home-manager/netrc.nix
@@ -13,6 +13,7 @@ in
   config = lib.mkIf (cfg.file != null) {
     home.activation = {
       linkNetrc = lib.hm.dag.entryAfter [ "writeBoundary" ] ''
+        [ -f ${config.age.secrets.netrc.path} ] || exit 0
         ln -sf ${config.age.secrets.netrc.path} ${config.home.homeDirectory}/.netrc
         chmod 644 ${config.home.homeDirectory}/.netrc
       '';
